+++ lib/queries/person/autocomplete.rb
@@ -24,7 +24,7 @@ module Queries
 
       # @return [Scope]
       def base_query
-        ::Person.select('people.*')
+        ::Person #.select('people.*')
       end
 
       # @return [Arel::Nodes::Equatity]
@@ -43,7 +43,7 @@ module Queries
       def autocomplete_exact_match
         base_query.where(
           table[:cached].eq(normalize_name).to_sql
-        ).limit(20)
+        ).limit(5)
       end
 
       def autocomplete_exact_last_name_match
@@ -56,7 +56,7 @@ module Queries
       def autocomplete_exact_inverted
         base_query.where(
           table[:cached].eq(invert_name).to_sql
-        ).limit(20)
+        ).limit(5)
       end
 
       def autocomplete_alternate_values_last_name
@@ -73,6 +73,7 @@ module Queries
       def normalize(string)
         n = string.strip.split(/\s*\,\s*/, 2).join(', ')
         n = 'nothing to match' unless n.include?(' ')
+        n
       end
 
       # @return [String] simple name inversion
@@ -120,16 +121,14 @@ module Queries
             a = q.left_outer_joins(:roles)
                   .joins("LEFT OUTER JOIN sources ON roles.role_object_id = sources.id AND roles.role_object_type = 'Source'")
                   .joins('LEFT OUTER JOIN project_sources ON sources.id = project_sources.source_id')
-                  .select('people.*, COUNT(roles.id) AS role_count, CASE WHEN MAX(roles.project_id) > 0 THEN MAX(roles.project_id) ELSE MAX(project_sources.project_id) END AS role_project_id')
-                  .where('roles.project_id = ? or roles.project_id IS NULL', Current.project_id)
-                  .where('project_sources.project_id = ? or project_sources.project_id IS NULL', Current.project_id)
-                  .group('people.id')
-                  .order('role_project_id, role_count DESC')
+                  .select('people.*, COUNT(roles.id) AS role_count, CASE WHEN MAX(roles.project_id) > 0 THEN MAX(roles.project_id) ELSE MAX(project_sources.project_id) END AS in_project_id')
+                  .where('roles.project_id = ? OR roles.project_id IS NULL OR project_sources.project_id = ? OR project_sources.project_id IS NULL', Current.project_id, Current.project_id)
+                  .group('people.id, roles.project_id')
+                  .order('in_project_id, role_count DESC')
           end
           a ||= q
           updated_queries[i] = a
         end
-
         result = []
         updated_queries.each do |q|
           result += q.to_a
@@ -141,12 +140,11 @@ module Queries
           result = ::Person.left_outer_joins(:roles)
                          .joins("LEFT OUTER JOIN sources ON roles.role_object_id = sources.id AND roles.role_object_type = 'Source'")
                          .joins('LEFT OUTER JOIN project_sources ON sources.id = project_sources.source_id')
-                         .select('people.*, COUNT(roles.id) AS role_count, CASE WHEN MAX(roles.project_id) > 0 THEN MAX(roles.project_id) ELSE MAX(project_sources.project_id) END AS role_project_id')
+                         .select('people.*, COUNT(roles.id) AS role_count, CASE WHEN MAX(roles.project_id) > 0 THEN MAX(roles.project_id) ELSE MAX(project_sources.project_id) END AS in_project_id')
                          .where('people.id IN (?)', result.collect{|i| i.id})
-                         .where('roles.project_id = ? or roles.project_id IS NULL', Current.project_id)
-                         .where('project_sources.project_id = ? or project_sources.project_id IS NULL', Current.project_id)
+                         .where('roles.project_id = ? OR roles.project_id IS NULL OR project_sources.project_id = ? OR project_sources.project_id IS NULL', Current.project_id, Current.project_id)
                          .group('people.id')
-                         .order('role_project_id, role_count DESC')
+                         .order('in_project_id, role_count DESC')
         end
       end