+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Cancel function for http requests
 
 ### Fixed
+- Updated sorting for people. People used in the same project are prioritized
 - Autocomplete in Interactive key task
 - Copy and clone option in Matrix Row Coder
 - Edit biological associations form, broken HTML and fields incorrectly mapped [#2370]